Manual typing up to 8.5.3, squashed
Extract stubs from rpy files in renpy/common/
Append file contents if rpym and rpy share the same name
Add typing to rpy files
Typing args and kwargs

+++ scripts/generate_pyi.py
@@ -1,3 +1,5 @@
+import ast
+import dataclasses
 import subprocess
 import sys
 import shutil
@@ -12,51 +14,86 @@ from .relative_imports import main as add_relative_imports
 ## STUB GENERATION
 
 
+@dataclasses.dataclass
+class StubFileContext:
+    source_python_files: list[str]
+    stubgen_output_path: Path
+
+
+def extract_py_from_rpy() -> None:
+    cmd = ["../renpy-8.5.0-sdk/lib/py3-windows-x86_64/python.exe", "./scripts/extract_py_from_rpy.py"]
+    _run_command(cmd)
+
+
 def generate_stubs() -> list[str]:
     """Generates .pyi stub files using stubgen and moves them to the target directory."""
-    _clean_temp_dir()
+    stub_file_contexts: list[StubFileContext] = []
 
-    created_files: list[str] = []
+    # Generate stubs for Ren'Py codebase
+    stub_file_contexts.append(
+        StubFileContext(
+            source_python_files=[str(config.renpy_path / "renpy")],
+            stubgen_output_path=config.temp_gen_path,
+        )
+    )
 
-    for source, target in config.renpy_directory_mapping.items():
-        source_path = config.renpy_path / source
-        target_path = config.pyi_path / target
+    # Generate stubs for extracted .rpy files
+    for dirpath, _, filenames in os.walk(config.rpy_extraction_temp_path):
+        extracted_name = Path(dirpath).relative_to(config.rpy_extraction_temp_path)
+        if len(extracted_name.parts) < 2 or extracted_name.parts[0] != "renpy":
+            continue
+        extracted_name = extracted_name.relative_to("renpy")
 
-        _run_command(["stubgen", "--include-private", "--output", str(config.temp_path), str(source_path)])
-        new_files = _move_stubs(source_path, target_path)
-        created_files.extend(new_files)
+        source_files = [os.path.join(dirpath, f) for f in filenames if f.endswith(".py")]
+        if not source_files:
+            continue
 
-    _clean_temp_dir()
+        stub_file_contexts.append(
+            StubFileContext(
+                source_python_files=source_files,
+                stubgen_output_path=config.temp_gen_path / "renpy" / extracted_name,
+            )
+        )
 
+    for ctx in stub_file_contexts:
+        _run_command([
+            "stubgen",
+            "--include-private",
+            "--output",
+            str(ctx.stubgen_output_path),
+            *ctx.source_python_files,
+        ])
+
+    created_files = _move_stubs(config.temp_gen_path / "renpy", config.pyi_path / "renpy-stubs")
     return created_files
 
 
 def _move_stubs(source_path: Path, target_path: Path) -> list[str]:
-    stubgen_output_path = config.temp_path / source_path.name
-
-    if not stubgen_output_path.exists():
+    if not source_path.exists():
         print(
-            f"[ERROR] stubgen did not create expected output directory: {stubgen_output_path}",
+            f"[ERROR] stubgen did not create expected output directory: {source_path}",
             file=sys.stderr,
         )
         sys.exit(4)
 
-    if target_path.exists():
-        if target_path.is_dir():
-            shutil.rmtree(target_path)
-        else:
-            os.remove(target_path)
-
     # Exclude specific files if needed
     for fn in config.exclude_files:
-        path_to_exclude = (config.temp_path / fn).with_suffix(".pyi")
-        if path_to_exclude.exists():
-            if path_to_exclude.is_dir():
-                shutil.rmtree(path_to_exclude)
-            else:
+        path_to_exclude = config.temp_gen_path / fn
+
+        if path_to_exclude.exists() and path_to_exclude.is_dir():
+            shutil.rmtree(path_to_exclude)
+        else:
+            path_to_exclude = path_to_exclude.with_suffix(".pyi")
+            if path_to_exclude.exists():
                 os.remove(path_to_exclude)
 
-    os.rename(str(stubgen_output_path), str(target_path))
+    shutil.copytree(str(source_path), str(target_path), dirs_exist_ok=True)
+    shutil.rmtree(str(source_path))
+
+    for p in target_path.glob("**/*.pyi"):
+        if os.path.getsize(p) == 0:
+            os.remove(str(p))
+
     return [str(p) for p in target_path.glob("**/*.pyi")]
 
 
@@ -71,7 +108,7 @@ def clean_generated_files(files: list[str]) -> None:
             with open(fn, "r", encoding="utf-8") as f:
                 lines = f.readlines()
 
-            with open(fn, "w", encoding="utf-8") as f:
+            with open(fn, "w", encoding="utf-8", newline="\n") as f:
                 for line in lines:
                     if line.startswith("from renpy.compat import"):
                         continue
@@ -121,13 +158,27 @@ def _run_command(cmd: list[str]) -> None:
         sys.exit(completed.returncode)
 
 
+def _remove_existing_directory():
+    path = config.pyi_path / "renpy"
+    if path.exists():
+        shutil.rmtree(path)
+
+
 def _clean_temp_dir():
-    if os.path.exists(config.temp_path):
-        shutil.rmtree(config.temp_path)
+    if os.path.exists(config.rpy_extraction_temp_path):
+        shutil.rmtree(config.rpy_extraction_temp_path)
+    if os.path.exists(config.temp_gen_path):
+        shutil.rmtree(config.temp_gen_path)
 
 
 def main():
+    _clean_temp_dir()
+    # _remove_existing_directory()
+
+    extract_py_from_rpy()
     created_files = generate_stubs()
+
+    _clean_temp_dir()
     clean_generated_files(created_files)
     add_relative_imports()
     format_code()
